+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Enhancement
 - Updated Google Omniauth to no longer write to disk
+- Improved memory usage for large sheet exports
 
 ## 0.27.1 (November 21, 2014)